Add files via upload
CV2 package automatically solve the negative values. They are displayed as a black([0,0,0]) area. In this case, nn.Softplus is not necessary.
Another possiblity is that: difference = figureA - figureB may not be solid method to detect foreground(in other tasks, it could be background). I need to check each pixel that was in the blue/green/red area.